[level 1] Title: x만큼 간격이 있는 n개의 숫자, Time: 0.21 ms, Memory: 3.8 MB -BaekjoonHub

+++ 프로그래머스/1/12954. x만큼 간격이 있는 n개의 숫자/x만큼 간격이 있는 n개의 숫자.cpp
@@ -7,7 +7,7 @@ vector<long long> solution(int x, int n) {
     vector<long long> answer;
     
     for (int i = 1 ; i <= n ; i++) {
-        answer.push_back(x*i);
+        answer.push_back((long long)x*i);
     }
     
     return answer;
